refactor(calPoker): own outcome projection
Keep CalPoker result parsing and card ordering inside the feature instead of making shared session types depend on a game-specific module.
