+++ front-end/src/components/GameSession.tsx
@@ -5,7 +5,8 @@ import { useCalpokerHand } from '../features/calPoker/useCalpokerHand';
 import { CalpokerDisplaySnapshot, SessionSave } from '../hooks/save';
 import { formatMojos, formatAmount } from '../util';
 import { getPlayerId } from '../hooks/save';
-import { CalpokerOutcome, SessionPhase } from '../types/ChiaGaming';
+import { SessionPhase } from '../types/ChiaGaming';
+import { CalpokerOutcome } from '../features/calPoker/outcome';
 import { SessionController, RestoreStatus } from '../hooks/SessionController';
 import type { BlockchainPoller } from '../hooks/BlockchainPoller';
 import Calpoker from '../features/calPoker';
