+++ front-end/src/types/ChiaGaming.ts
@@ -1,7 +1,6 @@
 import { CoinRecord } from './rpc/CoinRecord';
 import { Program } from 'clvm-lib';
 import { jsonStringify } from '../util/jsonSafe';
-import { cardIdToRankSuit } from '../features/calPoker/types/cardHelpers';
 
 export type HubLiveness = 'connected' | 'reconnecting' | 'inactive' | 'disconnected';
 
@@ -507,29 +506,6 @@ export interface WatchReport {
   deleted_watched: string[];
 }
 
-function select_cards_using_bits<T>(card: T[], mask: bigint): T[][] {
-  const result0: T[] = [];
-  const result1: T[] = [];
-  card.forEach((c, i) => {
-    if ((mask & (1n << BigInt(i))) !== 0n) {
-      result1.push(c);
-    } else {
-      result0.push(c);
-    }
-  });
-  return [result0, result1];
-}
-
-function compare_card(a: bigint, b: bigint): number {
-  const aRankSuit = cardIdToRankSuit(a);
-  const bRankSuit = cardIdToRankSuit(b);
-  const rankdiff = aRankSuit.rank - bRankSuit.rank;
-  if (rankdiff === 0) {
-    return aRankSuit.suit - bRankSuit.suit;
-  }
-  return rankdiff;
-}
-
 export interface PeerConnectionResult {
   /** Returns false when the hub WS is not OPEN (frame was not sent). */
   sendMessage: (msgno: number, input: Uint8Array) => boolean;
@@ -539,112 +515,6 @@ export interface PeerConnectionResult {
   sendKeepalive: () => boolean;
   hostLog: (msg: string) => void;
   close: () => void;
-}
-
-export class CalpokerOutcome {
-  alice_discards: bigint;
-  bob_discards: bigint;
-
-  alice_selects: bigint;
-  bob_selects: bigint;
-
-  alice_hand_value: bigint[];
-  bob_hand_value: bigint[];
-
-  win_direction: bigint;
-  my_win_outcome: 'win' | 'lose' | 'tie';
-
-  alice_cards: bigint[];
-  bob_cards: bigint[];
-
-  alice_final_hand: bigint[];
-  bob_final_hand: bigint[];
-
-  alice_used_cards: bigint[];
-  bob_used_cards: bigint[];
-
-  my_cards: bigint[];
-  their_cards: bigint[];
-  my_final_hand: bigint[];
-  their_final_hand: bigint[];
-  my_used_cards: bigint[];
-  their_used_cards: bigint[];
-  my_hand_value: bigint[];
-  their_hand_value: bigint[];
-
-  constructor(
-    iStarted: boolean,
-    myDiscards: bigint,
-    alice_cards: bigint[],
-    bob_cards: bigint[],
-    readableBytes: Uint8Array | number[],
-  ) {
-    const program = Program.deserialize(Uint8Array.from(readableBytes));
-    const result_list = program.toList();
-    this.alice_cards = alice_cards;
-    this.bob_cards = bob_cards;
-
-    this.alice_selects = result_list[1].toBigInt();
-    this.bob_selects = result_list[2].toBigInt();
-    this.alice_hand_value = result_list[3].toList().map(v => v.toBigInt());
-    this.bob_hand_value = result_list[4].toList().map(v => v.toBigInt());
-    let raw_win_direction = result_list[5].toBigInt();
-    if (iStarted) {
-      raw_win_direction *= -1n;
-      this.alice_discards = result_list[0].toBigInt();
-      this.bob_discards = myDiscards;
-    } else {
-      this.alice_discards = myDiscards;
-      this.bob_discards = result_list[0].toBigInt();
-    }
-
-    this.win_direction = raw_win_direction;
-    const alice_win = this.win_direction < 0n;
-
-    if (this.win_direction === 0n) {
-      this.my_win_outcome = 'tie';
-    } else if (alice_win) {
-      this.my_win_outcome = iStarted ? 'win' : 'lose';
-    } else {
-      this.my_win_outcome = iStarted ? 'lose' : 'win';
-    }
-
-    const [alice_for_alice, alice_for_bob] = select_cards_using_bits(
-      this.alice_cards,
-      this.alice_discards,
-    );
-    const [bob_for_bob, bob_for_alice] = select_cards_using_bits(
-      this.bob_cards,
-      this.bob_discards,
-    );
-
-    this.alice_final_hand = [...bob_for_alice];
-    alice_for_alice.forEach((c) => this.alice_final_hand.push(c));
-    this.alice_final_hand.sort(compare_card);
-
-    this.bob_final_hand = [...alice_for_bob];
-    bob_for_bob.forEach((c) => this.bob_final_hand.push(c));
-    this.bob_final_hand.sort(compare_card);
-
-    this.alice_used_cards = select_cards_using_bits(
-      this.alice_final_hand,
-      this.alice_selects,
-    )[1];
-    this.bob_used_cards = select_cards_using_bits(
-      this.bob_final_hand,
-      this.bob_selects,
-    )[1];
-
-    const iAmAlice = !iStarted;
-    this.my_cards = iAmAlice ? this.alice_cards : this.bob_cards;
-    this.their_cards = iAmAlice ? this.bob_cards : this.alice_cards;
-    this.my_final_hand = iAmAlice ? this.alice_final_hand : this.bob_final_hand;
-    this.their_final_hand = iAmAlice ? this.bob_final_hand : this.alice_final_hand;
-    this.my_used_cards = iAmAlice ? this.alice_used_cards : this.bob_used_cards;
-    this.their_used_cards = iAmAlice ? this.bob_used_cards : this.alice_used_cards;
-    this.my_hand_value = iAmAlice ? this.alice_hand_value : this.bob_hand_value;
-    this.their_hand_value = iAmAlice ? this.bob_hand_value : this.alice_hand_value;
-  }
 }
 
 export interface BlockchainReport {
